Add Delft fixture and reorganise tests by scenario
Canonical 4-file Delft fixture (zones.gpkg 1792 zones, depots.gpkg 16
depots, carrier_shares.csv 8 carriers, skim_time.mtx float32 sub-matrix)
extracted from mass-gt LEADVersion data. Zones include study area,
depot-only, and external COROP zones; households=0/employment=0 signals
no demand.

SkimMatrix.from_file gains a dtype parameter so float32 MASS-GT skims
load correctly without bit-reinterpretation.

Tests reorganised: joinville/ and delft/ each contain their own fixtures,
toml config, and test files. Old top-level test files removed.

README example output section updated with Delft choropleth and fixed
Joinville image path.

+++ src/urban_dollop/models/zone.py
@@ -20,8 +20,8 @@ class Zone(BaseModel):
     """
 
     zone_id: int
-    households: int
-    employment: int
+    households: float
+    employment: float
 
     @classmethod
     def from_file(cls, path: str | Path, columns: dict[str, str] = {}) -> list[Self]:
